Change the dummy terminal to a "raw" terminal that attempts to pass on input.
